refactor(ui): unify the link card into @kkhys/ui
me (horizontal card with description, favicon, and responsive blur-up
image) and memo (a diverged compact copy with its own blur mechanism,
hardcoded radius, and different border token) each maintained a full
link-card implementation. Extract me's presentation as a display-only
@kkhys/ui/link-card.astro taking metadata via props; each app keeps its
own fetch layer (me: validated metadata, memo: fetch-site-metadata with
dev fallback) behind a thin wrapper at the old path so consumers don't
move. The favicon resolver and the byte-signature sniffing behind it move
to the package next to the card, so memo resolves declared icons the same
way me does instead of always rendering the globe fallback. memo adopts
the unified look — its cards gain description and favicon rows.
blur-load handlers (and their test) move to the package since me's
content images share them.

Keep the guard memo's implementation had around placeholder generation:
an og:image that can't be fetched or decoded at build time degrades to no
blur-up rather than failing the build, which matters because memo deploys
unattended from CI.
