+++ apps/me/src/components/ui/image/image.astro
@@ -2,7 +2,7 @@
 import { Picture as AstroPicture, getImage, type Picture } from "astro:assets";
 import type { UnresolvedImageTransform } from "astro";
 
-import { blurLoadHandlers } from "#/utils/blur-load";
+import { blurLoadHandlers } from "@kkhys/ui/blur-load";
 import type { CustomImageProps } from "./types";
 
 type Props = CustomImageProps;
