+++ apps/memo/src/components/link-card.astro
@@ -1,5 +1,6 @@
 ---
-import { getImage, Image } from "astro:assets";
+import { type Favicon, getFavicon } from "@kkhys/ui/favicon";
+import LinkCard from "@kkhys/ui/link-card.astro";
 import type { HTMLAttributes } from "astro/types";
 import { getMetadata } from "#/lib/metadata";
 
@@ -7,112 +8,18 @@ interface Props extends HTMLAttributes<"a"> {
   href: string;
 }
 
-const { href, ...props } = Astro.props;
+const props = Astro.props;
 
-const { title, image } = await getMetadata(href);
+const { title, description, image, icon } = await getMetadata(props.href);
 
-let blurSrc: string | undefined;
-if (image) {
-  try {
-    const blurry = await getImage({
-      src: image.src,
-      width: 20,
-      format: "webp",
-    });
-    blurSrc = blurry.src;
-  } catch {
-    // Fall back to no blur if thumbnail generation fails
-  }
-}
+const favicon: Favicon = icon ? await getFavicon(icon) : { kind: "none" };
 ---
 
-<div class="link-card-container">
-  <a
-      href={href}
-      target="_blank"
-      rel="noopener noreferrer"
-      {...props}
-  >
-    {image && (
-      <div class="link-card-image-wrapper" data-blur-load>
-        <Image
-            src={image.src}
-            alt={`Cover image of ${title}`}
-            width={800}
-            height={420}
-            format="avif"
-            onload="this.closest('[data-blur-load]')?.classList.add('image-loaded')"
-        />
-      </div>
-    )}
-    <div class="link-card-content">
-      <span class="link-card-hostname">
-        {new URL(href).hostname}
-      </span>
-      <span class="link-card-title">
-        {title}
-      </span>
-    </div>
-  </a>
-</div>
-
-<style define:vars={{ 'blur-bg': blurSrc ? `url(${blurSrc})` : 'none' }}>
-  .link-card-container {
-    user-select: none;
-    overflow: hidden;
-    border-radius: 0.75rem;
-    border: 1px solid var(--c-border);
-    max-width: 400px;
-    transition: transform 0.1s;
-  }
-
-  .link-card-container:active {
-    transform: scale(0.98);
-  }
-
-  .link-card-image-wrapper {
-    background: var(--blur-bg) center / cover no-repeat var(--c-surface);
-    filter: blur(36px);
-    transition: filter 250ms ease-in-out;
-  }
-
-  .link-card-image-wrapper.image-loaded {
-    background: none;
-    filter: none;
-  }
-
-  .link-card-image-wrapper :global(img) {
-    display: block;
-    width: 100%;
-    height: auto;
-  }
-
-  .link-card-content {
-    padding: 0.75rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.4rem;
-  }
-
-  .link-card-image-wrapper + .link-card-content {
-    border-top: 1px solid var(--c-border);
-  }
-
-  .link-card-hostname {
-    font-size: var(--fs-xs);
-    line-height: var(--lh-xs);
-    color: var(--c-sub);
-  }
-
-  .link-card-title {
-    font-size: var(--fs-sm);
-    font-weight: var(--fw-semibold);
-    line-height: var(--lh-sm);
-    color: var(--c-text);
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    word-break: break-word;
-  }
-</style>
+<LinkCard
+  {...props}
+  title={title}
+  description={description}
+  imageSrc={image?.src}
+  favicon={favicon}
+  external
+/>
