+++ apps/memo/src/layouts/layout.astro
@@ -20,7 +20,7 @@ import "#/styles/global.css";
         <slot name="structured-data" />
     </head>
     <body>
-        <BlurLoadNoscript selector="[data-blur-load]" />
+        <BlurLoadNoscript selector="[data-blur-load], .blur-load" />
         <Header />
         <main>
             <slot />
