+++ packages/ui/package.json
@@ -10,7 +10,11 @@
     "./budoux": "./src/budoux.ts",
     "./site-header.astro": "./src/site-header.astro",
     "./site-footer.astro": "./src/site-footer.astro",
-    "./blur-load-noscript.astro": "./src/blur-load-noscript.astro"
+    "./blur-load-noscript.astro": "./src/blur-load-noscript.astro",
+    "./link-card.astro": "./src/link-card.astro",
+    "./blur-load": "./src/blur-load.ts",
+    "./favicon": "./src/favicon.ts",
+    "./image-signature": "./src/image-signature.ts"
   },
   "scripts": {
     "check": "tsc --noEmit",
