+++ packages/ui/src/__tests__/favicon.test.ts
@@ -11,19 +11,23 @@ const ICO_BYTES = new Uint8Array([0x00, 0x00, 0x01, 0x00, 0x01, 0x00, 0, 0, 0, 0
 
 const SVG_TEXT = '<svg xmlns="http://www.w3.org/2000/svg"><rect /></svg>';
 
-const mockFetch = (body: BodyInit | null, init?: ResponseInit) => {
+// Spelled from the Response constructor rather than DOM's BodyInit: this
+// package type-checks against bun's lib, which has no DOM globals.
+type ResponseBody = ConstructorParameters<typeof Response>[0];
+
+const mockFetch = (body: ResponseBody, init?: ResponseInit) => {
   const spy = vi.fn<typeof fetch>().mockResolvedValue(new Response(body, init));
   vi.stubGlobal("fetch", spy);
   return spy;
 };
 
 describe("getFavicon", () => {
-  let getFavicon: typeof import("#/lib/api/favicon").getFavicon;
+  let getFavicon: typeof import("../favicon").getFavicon;
 
   beforeEach(async () => {
     // Fresh module per test so the module-level cache doesn't leak between tests.
     vi.resetModules();
-    ({ getFavicon } = await import("#/lib/api/favicon"));
+    ({ getFavicon } = await import("../favicon"));
   });
 
   afterEach(() => {
@@ -31,14 +35,14 @@ describe("getFavicon", () => {
   });
 
   it("returns a remote favicon for bytes sharp can decode", async () => {
-    mockFetch(PNG_BYTES as unknown as BodyInit);
+    mockFetch(PNG_BYTES as unknown as ResponseBody);
 
     const result = await getFavicon("https://example.com/icon.png");
     expect(result).toEqual({ kind: "remote", src: "https://example.com/icon.png" });
   });
 
   it("upgrades an http favicon to https before fetching", async () => {
-    const spy = mockFetch(PNG_BYTES as unknown as BodyInit);
+    const spy = mockFetch(PNG_BYTES as unknown as ResponseBody);
 
     const result = await getFavicon("http://example.com/icon.png");
     expect(result).toEqual({ kind: "remote", src: "https://example.com/icon.png" });
@@ -46,7 +50,7 @@ describe("getFavicon", () => {
   });
 
   it("inlines an ICO favicon as a data URI", async () => {
-    mockFetch(ICO_BYTES as unknown as BodyInit);
+    mockFetch(ICO_BYTES as unknown as ResponseBody);
 
     const result = await getFavicon("https://example.com/favicon.ico");
     expect(result).toEqual({
@@ -104,7 +108,7 @@ describe("getFavicon", () => {
   it("rejects an ICO too large to inline", async () => {
     const huge = new Uint8Array(50 * 1024 + 1);
     huge.set(ICO_BYTES);
-    mockFetch(huge as unknown as BodyInit);
+    mockFetch(huge as unknown as ResponseBody);
 
     const result = await getFavicon("https://example.com/huge.ico");
     expect(result).toEqual({ kind: "none" });
@@ -113,7 +117,7 @@ describe("getFavicon", () => {
   it("keeps a raster favicon regardless of size", async () => {
     const huge = new Uint8Array(50 * 1024 + 1);
     huge.set(PNG_BYTES);
-    mockFetch(huge as unknown as BodyInit);
+    mockFetch(huge as unknown as ResponseBody);
 
     const result = await getFavicon("https://example.com/huge.png");
     expect(result.kind).toBe("remote");
@@ -127,7 +131,7 @@ describe("getFavicon", () => {
   });
 
   it("caches results per icon URL", async () => {
-    const spy = mockFetch(PNG_BYTES as unknown as BodyInit);
+    const spy = mockFetch(PNG_BYTES as unknown as ResponseBody);
 
     const first = await getFavicon("https://example.com/cached.png");
     const second = await getFavicon("https://example.com/cached.png");
