+++ packages/ui/src/favicon.ts
@@ -1,6 +1,5 @@
 import { Buffer } from "node:buffer";
-import { createResolvedCache } from "#/lib/api/cache";
-import { isIcoImage, isRasterImage, isSvgDocument } from "#/utils/image-signature";
+import { isIcoImage, isRasterImage, isSvgDocument } from "./image-signature";
 
 /**
  * Favicon resolved at build time for a link card.
@@ -19,7 +18,9 @@ const NONE: Favicon = { kind: "none" };
 // oversized multi-resolution .ico files fall back to the globe icon instead.
 const MAX_INLINE_BYTES = 50 * 1024;
 
-const cache = createResolvedCache<Favicon>();
+// One resolution per icon URL for the whole build: a single page can render
+// many cards pointing at the same host.
+const cache = new Map<string, Favicon>();
 
 // Astro only optimizes https remote images (`image.remotePatterns`), so an
 // http favicon would be emitted verbatim and blocked as mixed content.
@@ -34,23 +35,31 @@ const inline = (mime: string, bytes: Uint8Array): Favicon => ({
 // `fetch-site-metadata` guesses `/favicon.ico` without checking it exists
 // (suppressAdditionalRequest), and declared icons can 404 or serve HTML, so
 // verify the bytes here; anything dubious degrades to the globe fallback.
-export const getFavicon = (url: string): Promise<Favicon> =>
-  cache(url, async () => {
-    try {
-      const src = toHttps(url);
-      const response = await fetch(src, { headers: { accept: "image/*" } });
-      if (!response.ok) return NONE;
+const resolve = async (url: string): Promise<Favicon> => {
+  try {
+    const src = toHttps(url);
+    const response = await fetch(src, { headers: { accept: "image/*" } });
+    if (!response.ok) return NONE;
 
-      const bytes = new Uint8Array(await response.arrayBuffer());
-      if (bytes.length === 0) return NONE;
+    const bytes = new Uint8Array(await response.arrayBuffer());
+    if (bytes.length === 0) return NONE;
 
-      if (isRasterImage(bytes)) return { kind: "remote", src };
+    if (isRasterImage(bytes)) return { kind: "remote", src };
 
-      if (bytes.length > MAX_INLINE_BYTES) return NONE;
-      if (isIcoImage(bytes)) return inline("image/x-icon", bytes);
-      if (isSvgDocument(bytes)) return inline("image/svg+xml", bytes);
-      return NONE;
-    } catch {
-      return NONE;
-    }
-  });
+    if (bytes.length > MAX_INLINE_BYTES) return NONE;
+    if (isIcoImage(bytes)) return inline("image/x-icon", bytes);
+    if (isSvgDocument(bytes)) return inline("image/svg+xml", bytes);
+    return NONE;
+  } catch {
+    return NONE;
+  }
+};
+
+export const getFavicon = async (url: string): Promise<Favicon> => {
+  const cached = cache.get(url);
+  if (cached) return cached;
+
+  const favicon = await resolve(url);
+  cache.set(url, favicon);
+  return favicon;
+};
